refactor: move accept-terms from /preferences to /account
Terms acceptance is a legal acknowledgment, not a user preference.
/account/accept-terms is semantically correct alongside /account (DELETE).
